requirements.txt
streamlit
easyocr
Pillow
requests
opencv-python
matplotlib
pandas